+++ Package/test/Core Tests/Safety Tests/safety_plan_object_test.dart
@@ -1,0 +1,22 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:aureus/aureus.dart';
+
+void main() {
+  group('SafetyPlanObject', () {
+    test('throws an error if settings length is not less than 0', () {
+      // Arrange
+      final Map<SafetyPlanOptions, bool> testSettings = {
+        SafetyPlanOptions.disableBiometrics: true,
+        SafetyPlanOptions.disableScreenshots: false,
+      };
+
+      // Act & Assert
+      expect(
+        () => SafetyPlanObject(settings: testSettings),
+        throwsA(
+          isA<AssertionError>(),
+        ),
+      );
+    });
+  });
+}