+++ Package/test/Widget Tests/Element Widg Tests/Accessibility/universal_gesture_detector_test.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:aureus/src/Widgets/Elements/Accessibility/UniversalGestureDetectorElement.dart'; // Import the file where UniversalGestureDetector is defined
+
+void main() {
+  testWidgets('UniversalGestureDetector should trigger onDetect when tapped',
+      (WidgetTester tester) async {
+    bool detected = false;
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: UniversalGestureDetector(
+          onDetect: () {
+            detected = true;
+          },
+          child: Container(),
+        ),
+      ),
+    );
+
+    await tester.tap(find.byType(UniversalGestureDetector));
+    await tester.pump();
+
+    expect(detected, true);
+  });
+
+  testWidgets(
+      'UniversalGestureDetector should trigger onDetect when Enter key is pressed',
+      (WidgetTester tester) async {
+    bool detected = false;
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: UniversalGestureDetector(
+          onDetect: () {
+            detected = true;
+          },
+          child: Container(),
+        ),
+      ),
+    );
+
+    await tester.sendKeyEvent(LogicalKeyboardKey.enter);
+    await tester.pump();
+
+    expect(detected, true);
+  });
+}